[Server|Sensor] Sensor 데이터 생산자 소비자 패턴 테스트 2

+++ Server/src/controllers/SensorDataQueue.cpp
@@ -24,7 +24,7 @@ void SensorDataQueue::destroyInstance()
     }
 }
 
-void SensorDataQueue::enqueue(id_t plantId, double temperature, double humidity)
+void SensorDataQueue::enqueue(id_t plantId, double temperature, int humidity)
 {
     QMutexLocker locker(&mutex_);
